// src/definitions/enums/document.enum.ts
export enum SearchDocumentOutTitles {
  UPDATEDATE = "UPDATEDATE",
  CREATEDATE = "CREATEDATE",
  NUMBER_ARRIVAL = "NUMBER_ARRIVAL",
  DATE_ARRIVAL = "DATE_ARRIVAL",
  DATE_RECEIVED = "DATE_RECEIVED",
  DEADLINE = "DEADLINE",
  DOCID = "DOCID",
  NUMBERSIGN = "NUMBERSIGN",
  PREVIEW = "PREVIEW",
  ORG_ISSUED_NAME = "ORG_ISSUED_NAME",
  ORG_HANDLE = "ORG_HANDLE",
  DOCTYPE = "DOCTYPE",
  DATEISSUED = "DATEISSUED",
  PERSON_HANDLE = "PERSON_HANDLE",
  SECURITY_NAME = "SECURITY_NAME",
  STATUS = "STATUS",
  IMPORTANT = "IMPORTANT",
  ORGTRANSFER = "ORGTRANSFER",
}

export enum OBJ_TYPE {
  VAN_BAN_DEN,
  VAN_BAN_DI,
  GIAO_VIEC,
  VAN_BAN_NOI_BO,
  LICH,
  VAN_BAN_XIN_XE,
}
export enum CERT_OBJ_TYPE {
  doc_in_transfer = "doc_in_transfer",
  doc_in_return = "doc_in_return",
  doc_in_done = "doc_in_done",
  doc_in_org_transfer = "doc_in_org_transfer",
  doc_out_internal = "doc_out_internal",
  doc_in_add = "doc_in_add",
  doc_in_comment = "doc_in_comment",
  doc_out_add = "doc_out_add",
  doc_out_comment = "doc_out_comment",
  doc_out_done = "doc_out_done",
  task = "task",
  task2 = "task2",
  task_comment = "task_comment",
  task_result = "task_result",
  org = "org",
  user = "user",
  doc_internal = "doc_internal",
  doc_internal_comment = "doc_internal_comment",
  calendar = "calendar",
}
export enum TabNames {
  DUTHAO = "draft",
  DATRINHKY = "draftOnSign",
  DABANHANH = "draftIssued",
  CHOXULY = "waitHandleTab",
  CHOCHOYKIEN = "waitCommentTab",
  DAXULY = "handledTab",
  DABANHANH_DOC = "issued",
  CHOBANHANH = "waitIssued",
}